Moved functions to deman_acep and modified test script
