+++ demand_acep/read_NetCDF.py
@@ -2,13 +2,16 @@
 import os
 import pandas as pd
 import pdb
-import xarray as xr
 
 
+from demand_acep import extract_data
+from demand_acep import extract_ppty
 # %% Paths
 path = os.getcwd()
-path_ppty = os.path.join(path, 'demand_acep/data/properties')
-path_data = os.path.join(path, 'demand_acep/data/measurements')
+# path_ppty = os.path.join(path, 'demand_acep/data/properties')
+# path_data = os.path.join(path, 'demand_acep/data/measurements')
+path_ppty = os.path.join(path, 'data/properties')
+path_data = os.path.join(path, 'data/measurements')
 # %% Read in files containing data type
 filename_ppty = 'Copy of Measured Channels PFRR.xlsx'
 meter_details = pd.read_excel(os.path.join(path_ppty, filename_ppty))
@@ -18,32 +21,6 @@ channel_description = list(meter_details['Desc'][:48])
 channel_dict = dict(zip(channel_name, channel_description))
 # Extract name of meters
 meter_name = list(meter_details.columns.values)[-4:]
-# %% Functions
-
-
-def extract_data(dirpath, filename):
-    """This function takes in a filename and directory path, extracts the meter channel data in that file and
-    returns the time and channel values at each time"""
-    netcdf_data = xr.open_dataset(os.path.join(dirpath, filename))
-    netcdf_df = netcdf_data.to_dataframe()
-
-    return netcdf_df.index, netcdf_df['value'].values
-
-
-def extract_ppty(filename, meter_name):
-    """This function takes in a filename and the list containing the names of each of the four meters at pokerflats;
-    extracts and returns the meter name and measurement type described in the filename"""
-    filename_split_1 = filename.split('@')
-    filename_split_2 = filename_split_1[0].split('-')
-    meter_channel = filename_split_2[-1]
-    for name in meter_name:
-        if meter_channel.startswith(name):
-            n_name = len(name)
-            meter = meter_channel[:n_name]
-            channel = meter_channel[n_name:]
-
-    return meter, channel
-
 # %% Walk through folders and read .nc files using xarray
